Added favorites page

+++ src/components/DropDownItem.js
@@ -2,9 +2,9 @@ import React from 'react'
 import { useGoogleLogout } from 'react-google-login'
 import './DropDownItem.scss'
 import { TimelineLite, Power3 } from 'gsap'
+import { withRouter } from 'react-router-dom'
 
-
-function DropDownItem ({ text, type, reload }) {
+function DropDownItem ({ text, type, reload, history }) {
   const className = 'DropDownItem ' + type
 
   const { signOut } = useGoogleLogout({
@@ -14,21 +14,29 @@ function DropDownItem ({ text, type, reload }) {
     onLogoutSuccess: () => {console.log('Logout Success')}
   })
 
-  const hc = () => {
-    localStorage.clear()
-    reload()
-    const tl = new TimelineLite()
+  const handleClick = () => {
+    if (type === 'logout') {
+      localStorage.clear()
+      reload()
+      const tl = new TimelineLite()
 
-    tl.to('#NavMyAccount', {duration: 0.23,opacity: 0, ease: Power3.easeOut})
-    .set('#NavLogIn, #NavSignUp', {display: 'flex'})
-    .set('#NavMyAccount', {display: 'none'})
-    .to('#NavLogIn, #NavSignUp', {duration: 0.23, opacity: 1, ease: Power3.easeOut})
+      tl.to('#NavMyAccount', {duration: 0.23,opacity: 0, ease: Power3.easeOut})
+      .set('#NavLogIn, #NavSignUp', {display: 'flex'})
+      .set('#NavMyAccount', {display: 'none'})
+      .to('#NavLogIn, #NavSignUp', {duration: 0.23, opacity: 1, ease: Power3.easeOut})
     
-    signOut()
+      signOut()
+    } else if (type === 'favorites') {    
+      console.log('Favorites')
+      history.push({
+      pathname: '/favorites',
+      })
+    }
   }
+
   return (
-    <p className={className} onClick={hc}>{text}</p>
+    <p className={className} onClick={handleClick}>{text}</p>
   )
 }
 
-export default DropDownItem
+export default withRouter(DropDownItem)
